iio: adc: ti-ads1015: setup irq as IRQF_SHARED
It is possible to have multiple ADS1x15 alert/rdy lines wired to the
same irq: setup the IRQ as shared.

The ADC does not have a register to check interrupt status so
work-around by checking if value is in range.
This is approximate because in traditional modes a comparator between
low and high threshold should only trigger an alert if it comes down
from high, but this is good enough for now.

Signed-off-by: Syunya Ohshio <syunya.ohshio@atmark-techno.com>
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>
(cherry picked from commit da5829e9503d41806fd42070cb8d039065c3e4f4 @ https://github.com/atmark-techno/linux-5.10-at)
Signed-off-by: Daisuke Mizobuchi <mizo@atmark-techno.com>