LF-14260-6 net: enetc: add Frame Modification Table support
Ingress and egress frame modification actions are specified in this
table if Frame Modification entry ID (FM_EID) is used as an index.
This table is implemented as a linear array of words, accessed using
an index (0, 1, 2, ..., n) that uniquely identifies a word within the
array.  For this table, each table entry occupies one word (one to one
corresponding between words and entries).

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>