LF-14260-26 net: dsa: netc: add PTP support
The switch supports both one-step and two-step timestamp. For ingress
PTP packets, the filtering rules are set to redirect them to the CPU
port via the ingress port filter table. And the capture timestamp is
recorded in the switch tag (To_Host tag, subType=1).

For egress PTP packets, the switch tag(To_Port tag) is used to specify
whether to perform one-step timestamping (subType=1)or two-step
timestamping (subType=2). If two-step timestamping is performed, the
switch automatically generates a response packet to record the transmit
timestamp in the switch tag (To_Host tag, subType=2) and sends this
response packet to the CPU port.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>