LF-14260-28 net: dsa: netc: add debugfs interface to dump drop counters
The statistics of each user port can be obtained through ethtool,
but the statistics of the CPU port cannot be obtained by ethtoll.
Besides, ethtool obtains many statistical values, but only a few
key statistical values are needed in the actual debugging process.
To facilitate debugging, the debugfs interface is added to obtain
key statistical information of the specified port.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>