arm64: proton-pack: Add Spectre-BSE mitigation for Cortex-A7{2,3,5}
A group of older Cortex cores are affected by Spectre-BSE which on one
core bypasses the existing Spectre-BHB mitigation.

The workaround is to use the Spectre-BHB 'arch_workaround_3' firmware
call. For Cortex-A7{3,5}, this is the existing mitigation so only the
reporting needs updating.
For Cortex-A72 if WA3 is not implemented, there is still a benefit
from enabling the Spectre-BHB branchy-loop.

This leads to special-casing anything affected by Spectre-BSE to avoid
enabling the branchy-loop, and instead only do this if the firmware WA3
call is not implemented. Such platforms will be reported as:
| Mitigation: Branch predictor hardening, BHB, but not BSE
via the Spectre-v2 vulnerabilities file.

Signed-off-by: James Morse <james.morse@arm.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>