LF-14260-17 ptp: netc: get current time from TMR_CUR_TIME registers
TMR_CUR_TIME registers are used to get the current time, the time is
calculated by adding TMROFF_H/L with TMR_CNT_H/L register. Currently,
the TMROFF_H/L register is not used and its default value is 0, so
the value read from TMR_CNT registers is the same as TMR_CUR_TIME_H/L.
However, in order to prevent the potential issue caused by the use of
the TMROFF_H/L register in the future, it is best to use TMR_CUR_TIME
registers to obtain the current time from now on.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>