LF-14261-2 net: enetc: remove emdio consumer-supplier related interfaces
Some platforms like S32 do not use netc-blk-ctrl, therefore, the current
method will cause problems on these platforms. In addition, the internal
ENETC port of i.MX94 which connects to the CPU port of the switch also
does not depends on the EMDIO, so remove the relevant interfaces. A more
generic method will be used in the subsequent patch.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>