LF-14261-8 net: enetc: set the command cache attribute for i.MX ENETC SI
We have set SI's buffer cache attribute for Tx/Rx BD ring, but forgot to
set the cache attribute for the command BD ring. The command BD ring is
similar to Tx/Rx BD ring, on the LS1028A platform, the value of command
cache attribute is the same as that of buffer cache attribute. To keep
it consistent, it is also set to the same value on the i.MX platforms.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>