LF-14857-7 net: enetc: don't interrupt Rx reception due to bad packet
Previously, the Rx reception would be interrupted when a bad packet was
received. If there were still packets on the Rx BD ring, they could only
be received at the next interrupt, which could degrade performance and
increase latency. So it is better to continue receiving packets on the
Rx BD ring after processing the bad packet.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Acked-by: Clark Wang <xiaoning.wang@nxp.com>