LF-14282-3 net: enetc: save parsed PTP info in skb->cb
Previously, Tx PTP messages were parsed twice in the enetc driver, the
first time in enetc_xmit() and the second time in enetc_map_tx_buffs().
This is completely unnecessary because the parsed information can be
saved in skb->cb, so that enetc_map_tx_buffs() can get the previously
parsed data through skb->cb. Therefore, struct enetc_skb_cb is added as
the format of the data in the skb->cb buffer.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>