LF-14261-11 net: enetc: fix memory leak when doing suspend/resume test
Each time a suspend/resume test is performed, each enetc will leak 12KB
of memory. The root cause is that enetc_init_cbdr() is called during
resume to re-alloc for memory and initialize the Command BD ring, but
the memory of the Command BD ring is not freed during suspend. So add
enetc_free_cbdr() to enetc4_pf_power_down().

Fixes: 58fd162d4015 ("LF-11880-2 net: enetc: Add power off suspend/resume and WOL supports")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>