LF-14282-5 net: enetc: fix PTP one-step issue for ENETC v4
Previously the PTP one-step feature of i.MX95 and i.MX94 did not work at
all, because i.MX95 and i.MX94 hardware does not support "dma-coherent".
Therefore, the PTP synchronization packet must be modified before calling
dma_map_single() to set the DMA direction to DMA_TO_DEVICE. Otherwise,
the modification of the packet will be invalid, and the originTimestamp
and correction fields of the sent packet will still be the values before
the modification.

Fixes: 1ffd37d8a30b ("LF-10640-7 net: enetc: Add i.MX95 ENETC support")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>