LF-14844 clk: imx95-blk-ctl: Cache registers when RPM routines are called
If runtime PM is used for the clock providers and they're part of a
power domain, then the power domain supply will be cut off when runtime
suspended. That means all BLK CTL registers belonging to that power
domain will be reset. Hence, the clock settings will revert to default
values messing up the consumer clock settings.

Also, fix the suspend/resume routines as well, as the clock was left ON
when going to suspend.

Fixes: 5224b189462f ("clk: imx: add i.MX95 BLK CTL clk driver")
Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>