arm64: imx: fix linkage issue when ARCH_MXC=n
Compiling a .config obtained in the following way for arm64:

- make defconfig lsdk.config
- edit .config to set CONFIG_ARCH_MXC=n (because
  drivers/mxc/gpu-viv/hal/os/linux/kernel/gc_hal_kernel_os.c fails to
  build with W=1 C=1)
- make -j 8 Image.gz dtbs W=1 C=1

results in this linkage issue:

  ld.lld: error: undefined symbol: request_bus_freq
  >>> referenced by smc_abi.c:940 (drivers/tee/optee/smc_abi.c:940)
  >>>               drivers/tee/optee/smc_abi.o:(optee_smc_do_call_with_arg) in archive vmlinux.a
  >>> referenced by sdhci-esdhc-imx.c:1859 (drivers/mmc/host/sdhci-esdhc-imx.c:1859)
  >>>               drivers/mmc/host/sdhci-esdhc-imx.o:(sdhci_esdhc_imx_probe) in archive vmlinux.a
  >>> referenced by sdhci-esdhc-imx.c:2188 (drivers/mmc/host/sdhci-esdhc-imx.c:2188)
  >>>               drivers/mmc/host/sdhci-esdhc-imx.o:(sdhci_esdhc_runtime_resume) in archive vmlinux.a
  >>> referenced 5 more times

  ld.lld: error: undefined symbol: release_bus_freq
  >>> referenced by smc_abi.c:978 (drivers/tee/optee/smc_abi.c:978)
  >>>               drivers/tee/optee/smc_abi.o:(optee_smc_do_call_with_arg) in archive vmlinux.a
  >>> referenced by sdhci-esdhc-imx.c:2005 (drivers/mmc/host/sdhci-esdhc-imx.c:2005)
  >>>               drivers/mmc/host/sdhci-esdhc-imx.o:(sdhci_esdhc_imx_probe) in archive vmlinux.a
  >>> referenced by sdhci-esdhc-imx.c:2033 (drivers/mmc/host/sdhci-esdhc-imx.c:2033)
  >>>               drivers/mmc/host/sdhci-esdhc-imx.o:(sdhci_esdhc_imx_remove) in archive vmlinux.a
  >>> referenced 9 more times
  make[3]: *** [../scripts/Makefile.vmlinux:34: vmlinux] Error 1
  make[2]: *** [Makefile:1166: vmlinux] Error 2
  make[1]: *** [Makefile:224: __sub-make] Error 2

It can be explained as follows:

request_bus_freq() and release_bus_freq() are symbols provided by
arch/arm/mach-imx/busfreq-imx.c, object file which is built when
CONFIG_ARCH_MXC is enabled. Additionally, the symbols are defined and
exported when CONFIG_HAVE_IMX_BUSFREQ is enabled. The
include/linux/busfreq-imx.h header has shim definitions for the
situation where CONFIG_HAVE_IMX_BUSFREQ is disabled, allowing for
drivers to be theoretically successfully built when that option is
disabled.

Thus, logically, CONFIG_HAVE_IMX_BUSFREQ depends on CONFIG_ARCH_MXC.
And for the 32-bit arm architecture, that dependency is actually
expressed in arch/arm/mach-imx/Kconfig. But for arm64 it isn't:
CONFIG_HAVE_IMX_BUSFREQ symbol, which is user-selectable and part of
arch/arm64/Kconfig.platforms, and which incorrectly does _not_ depend on
CONFIG_ARCH_MXC. AFAICS, this second definition shouldn't exist, so
delete it.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>