gpio: imx-rpmsg: add suspend/resume to allow pins to be used as wakeup
This is an heresy from the linux gpio subsystem point of view,
but allow setting up wakeup on a gpio without a separate driver.

This is useful to enable wakeup while the user can still use the
gpio as input gpio separately.

As a condition for using the target pin as a wakeup, the target pin must
be unoccupied by any process at the time of preparing suspend. This is
because this GPIO controller uses its own pin via gpiolib API to prevent
simultaneous access from multiple processes. Also need to add
'wakeup-source' on the same node as the GPIO controller device.

Which pin is the source for wakeup with which interrupt type can be set
at runtime by writing to /sys/devices/... /setwake.

Signed-off-by: Shota Shimoyama <shota.shimoyama@atmark-techno.com>
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>
(cherry picked from commit bb2e378a48047bd77d84711390638fb39f1b62a7 @ https://github.com/atmark-techno/linux-5.10-at)