LF-14949 clk: clk-imx95-blk-ctl: Change i.MX94 LVDS clock parent
While testing multiple modes, some of them didn't show up on screen and
the display showed an "Invalid timing" error. It turned out that the
dispmix LVDS clock gate parent was improperly set to the /7 divider and
when choosing modes with a lower pixel clock the LVDS clock could not be
changed. According to the RM the LVDS phy clock should come directly from
LDB PLL.

Changing the parent clock fixed the issue and all modes show correctly.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>