drivers: firmware: imx: fix build with CONFIG_IMX_ELE_TRNG=m
It is possible to select a configuration where CONFIG_IMX_SEC_ENCLAVE=y
and CONFIG_IMX_ELE_TRNG=m. In that case, the kernel fails to link at the
modules creation stage as follows:

aarch64-buildroot-linux-gnu-ld: drivers/firmware/imx/se_ctrl.o:(.data+0x288): undefined reference to `ele_trng_init'
aarch64-buildroot-linux-gnu-ld: drivers/firmware/imx/se_ctrl.o:(.data+0x2e8): undefined reference to `ele_trng_init'

There are many problems with the current logic. Code built into the
kernel (se_ctrl.c) cannot reference symbols from code built as a module
(ele_trng.c). Additionally, the referenced module may be disabled in
Kconfig, and there needs to be compile-time fallback for that case as
well, as well as the option exists.

It just doesn't make sense to allow CONFIG_IMX_ELE_TRNG to be a module,
since ele_trng.c clearly was not written with the mandatory
MODULE_LICENSE(), etc specific to modules. Instead, it appears to have
been intended that the true random number generator support is an
optional part of the secure enclave driver.

To fix the bug, limit CONFIG_IMX_ELE_TRNG to just yes or no. Include the
TRNG support in the secure enclave driver when set to yes, and define
shims that evaluate ele_trng_init() to a NULL pointer when set to no.
The driver is prepared to handle the optional init_trng() function
pointer, and for imx93, imx94 and im95, when the TRNG support is
compiled out, they behave as the platforms that don't support it at all,
like imx8dxl.

Fixes: 43b03b264fbe ("LF-13910: drivers: firmware: imx: add support for imx93 soc")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>