LF-15007 ptp: netc: add external trigger-stamp support
The NETC Timer is capable of recording the timestamp on receipt of an
external pulse on a GPIO pin. For i.MX95, it supports two such external
triggers. The recorded value is saved in a 16 entry FIFO accessed by
TMR_ETTSa_H/L. An interrupt can be generated when the trigger occurs,
when the FIFO reaches a threshold, and if the FIFO overflows.

Signed-off-by: F.S. Peng <fushi.peng@nxp.com>
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>