LF-14993 iio: adc: load calibrated values even calibration failed
ADC calibration might fail because of the noise of reference voltage.
To avoid calibration fail, need to meet the following requirement:
ADC reference voltage Noise < 1.8V * 1/2^ENOB

For the case which the ADC reference voltage on board do not meet
the requirement, still load the calibrated values, so ADC can also
work but not that accurate.

Tested-by: Winteri Wang <dongjie.wang@nxp.com>
Signed-off-by: Haibo Chen <haibo.chen@nxp.com>