LF-15037 net: enetc: correct query/update actions when adding ISIT entry
When adding an ISIT entry, CFGEU of UPDATE_ACTIONS should be set and
QUERY_ACTIONS should be set to 0x1 (Query ENTRY_ID only). But in the
previous implementation, the 'qa' and 'ua' parameters of ntmp_fill_crd()
in ntmp_isit_add_or_update_entry() were reversed, resulting in CFGEU not
being set.

Fixes: d7cf1818fa0c ("LF-14261-12 net: enetc: refine NTMP driver for upstream")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>