LF-13949-3: media: imx8-isi: Handle HDMI RX device in get_frame_desc
Currently, the imx8-isi driver is using get_frame_desc in order to
identify the CSI virtual channels based on the connected streams.
But, the frame_desc feature is only specific to CSI protocol and the ISI
module can also receive data from non-CSI sources, like the HDMI
receiver.

This patch fixes the get_frame_desc usage so it will handle the HDMI
receiver which has a function of MEDIA_ENT_F_IO_DTV. All other function
types are not changed in this patch.

Signed-off-by: Oliver F. Brown <oliver.brown@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>