LF-15091-2 media: i2c: ox03c10: Implement workaround for the VS issue
The OX03C10 sensors seem to mess up the captured frames if the VS
exposure was set to a value higher than 4 before streaming, even though
the maximum allowed value is 31. To avoid this, do not allow user to set
VS higher than 4 when we're not streaming and, after streaming, lower VS
to 4 if it was set higher during streaming.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>