LF-15098-2 ptp: netc: add debugfs support to loopback pulse signal
The NETC Timer supports to loop back the output pulse signal of Fiper-n
into Trigger-n input, so that we can leverage this feature to validate
some other features without external hardware support. For example, we
can use it to test external trigger stamp (EXTTS). And we can combine
EXTTS with loopback mode to check whether the generation time of PPS is
aligned with an integral second of PHC, or the periodic output signal
(PTP_CLK_REQ_PEROUT) whether is generated at the specified time.

Therefore, add the debugfs interfaces to enable the loopback mode of
Fiper1 and Fiper2. An example command is as follows.

echo 1 > /sys/kernel/debug/netc_timer0/fiper1-loopback

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>