LF-15098-3 ptp: netc: make PPS generation time align with integral seconds
PPS is a signal generated every time the PHC reaches 1-rounded second,
but we found that PPS signal was output immediately after being enabled,
so the generation time is uncertain, we can easily reproduce the problem
with the following commands. That is, by using the loopback mode and the
external trigger stamp (EXTTS) feature to make the hardware capture the
timestamp when it receives the loopback PPS signal, the timestamp should
be almost equal to the time when the PPS pulse is generated.

$ echo 1 > /sys/kernel/debug/netc_timer0/fiper1-loopback
$ echo 0 1 > /sys/class/ptp/ptp0/extts_enable
$ echo 1 > /sys/class/ptp/ptp0/pps_enable
$ ./testptp -d /dev/ptp0 -e 1
event index 0 at 1744979995.916574603

$ echo 0 > /sys/class/ptp/ptp0/pps_enable
$ echo 1 > /sys/class/ptp/ptp0/pps_enable
$ ./testptp -d /dev/ptp0 -e 1
event index 0 at 1744980015.756272247

In order to make the PPS pulse be generated when PHC reaches 1-rounded
second, the alarm trigger is used in FS mode to insrtuct the fiper to
start outputing the PPS signal at an integral second. However, we found
that after the modification, the PPS generation time still deviates from
integral seconds by 200~300us, see below log.

$ ./testptp -d /dev/ptp0 -e 2
event index 0 at 1744984252.000210835
event index 0 at 1744984253.000325209

After further debugging and confirmation with the design team, since the
FIPER and ALARM are tied to GCLK, which is configured via PRSC, we will
inherently get increased uncertainty for larger values of PRSC. So they
recommend setting the PRSC to the default value of 2, as well as setting
PG=0 for both ALARM and FIPER. Then we can see the deviation dropped to
17ns, which is much better than before.

$ ./testptp -d /dev/ptp0 -e 1
event index 0 at 1745029616.000000017

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>