LF-15098-5 ptp: netc: make PEROUT pulse output at the specified time
When using loopback mode and EXTTS to measure the generation time of the
PTP_CLK_REQ_PEROUT pulse, the pulse generation time is always one pulse
cycle later than the specified time. The issue can be reproduced through
the following command, the specified time is 1745030980s, pulse cycle is
1s, and we can see the pulse generation time is at 1745030981s, which is
1s later than the specified time.

$ echo 1 > /sys/kernel/debug/netc_timer0/fiper1-loopback
$ echo 0 1 > /sys/class/ptp/ptp0/extts_enable
$ echo 0 1745030980 0 1 0 > /sys/class/ptp/ptp0/period
$ ./testptp -d /dev/ptp0 -e 1
event index 0 at 1745030981.000000016

After checking with the design team, they clarified that the first FIPER
pulse is ALARM + FIPER_INTERVAL + D, where D is an internal delay, so the
timing of the first pulse is not immediately upon passing ALARM. So the
ALARM should be one pulse period smaller than the specified time. After
correcting the ALARM value, we can see that the pulse is output at the
specified time.

$ echo 1 > /sys/kernel/debug/netc_timer0/fiper1-loopback
$ echo 0 1 > /sys/class/ptp/ptp0/extts_enable
$ echo 0 1745032380 0 1 0 > /sys/class/ptp/ptp0/period
$ ./testptp -d /dev/ptp0 -e 1
event index 0 at 1745032380.000000015

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>