LF-15098-1 ptp: netc: add "nxp,pps-channel" property to select PPS channel
The NETC Timer has three channels to generate PPS signal, however, the
PPS interface provided by the kernel does not provide parameter for
setting the PPS signal channel. Previously, we hard-coded the channel
to FIPER1 in the driver. However, this has limitations, because the PPS
pins are shared with other devices, the pin of FIPER1 on some boards may
be used by other devices. Therefore, for these boards, they use another
PPS channel to output the PPS signal. So the "nxp,pps-channel" property
is added to indicate which channel the current board uses to output the
PPS signal.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>