LF-15098-6 ptp: netc: rearm ALARM of PPS if there is a time drift
If there is a time drift when PPS is enabled, the PPS signal won't be
generated at an integral second of PHC. For example, the issue can be
reproduced by following commands.

Step 1: Enable the loopback mode of Fiper1, EXTTS and PPS, and get the
external time stamp.
$ echo 1 > /sys/kernel/debug/netc_timer0/fiper1-loopback
$ echo 0 1 > /sys/class/ptp/ptp0/extts_enable
$ echo 1 > /sys/class/ptp/ptp0/pps_enable
$ ./testptp -d /dev/ptp0 -e 1
event index 0 at 1745035220.000000016

Step 2: Shift the time of PHC by 10000000ns through testptp and then get
the external time stamp.
$ ./testptp -d /dev/ptp0 -n 10000000
$ ./testptp -d /dev/ptp0 -e 1
event index 0 at 1745035236.010000015

From the above results, we can see the deviation is increased from 16ns
to 10000015ns. To solve this issue, we need to rearm ALARM to make the
next PPS signal be generated at an integral second of PHC. The result
after modification is shown below, the deviation is 16ns, which is as
expected.

$ ./testptp -d /dev/ptp0 -n 10000000
$ ./testptp -d /dev/ptp0 -e 1
event index 0 at 1745322531.000000016

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>