LF-15098-8 ptp: netc: each FIPER uses a different ALARM
NETC Timer has 3 FIPERs and 2 ALRAMs. Previously, all FIPERs use ALARM1
by default. There is no problem if only one FIPER is used. However, there
will be problems if multiple FIPERs are used simultaneously. For example,
FIPER1 is used to output PPS signal, which is generated when PHC reaches
1-rounded second. Meanwhile, if FIPER2 is used to output PEROUT signal,
and the start time of PEROUT signal is not an integer second, then it
will cause the PPS signal to be re-output when PHC reaches the specified
time. Therefore, the PPS signal won't be aligned with an integer second.

So the best practice is to use different ALARMs for different FIPERs,
but there are some limitations, for example, we cannot use all three
FIPERs at the same time because there are only two ALARMs.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>