LF-15098-9 ptp: netc: rearm ALARM of PEROUT if there is a time drift
Similar to PPS, if there is a time drift when PEROUT starts outputting
pulse, this will cause the pulse generation time to deviate greatly from
the expected time. For example, the issue can be reproduced by following
commands.

Step 1: Enable the loopback mode of FIPER1, EXTTS and PEROUT, and then
get the external time stamp.
$ echo 1 > /sys/kernel/debug/netc_timer0/fiper1-loopback
$ echo 0 1 > /sys/class/ptp/ptp0/extts_enable
$ echo 0 100 0 1 0 > /sys/class/ptp/ptp0/period
$ ./testptp -d /dev/ptp0 -e 1
event index 0 at 100.000000016

Step 2: Shift the time of PHC by 10000000ns through testptp and then get
the external time stamp.
$ ./testptp -d /dev/ptp0 -n 10000000
$ ./testptp -d /dev/ptp0 -e 1
event index 0 at 120.010000014

From the above results, we can see the deviation is increased from 16ns
to 10000014ns. To solve this issue, we need to rearm ALARM to make the
next pulse signal be generated at an integral second of PHC. The result
after modification is shown below, the deviation is 16ns which is as
expected.

$ ./testptp -d /dev/ptp0 -n 10000000
$ ./testptp -d /dev/ptp0 -e
event index 0 at 132.000000016

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>